Crate aqua_test_utils

Macros

call_vm

Structs

AquamarineVM

AquamarineVMConfig
Describes behaviour of the Aquamarine VM stepper.

NEVec
NEVec<T> represents a non-empty Vec<T>. It derefs to Vec<T> directly.

ParticleParameters
Information about the particle that is being executed by the stepper at the moment

StepperOutcome
Describes a result returned at the end of the stepper execution.

Enums

AquamarineVMError

IType
Represents the types supported by WIT.

IValue
A WIT value.

Functions

create_aqua_vm

echo_number_call_service

echo_string_call_service

set_variable_call_service

set_variables_call_service

unit_call_service

Type Definitions

CallServiceClosure